// apps/backend/src/app/error/enums/errorCodeEnum.ts
export enum ErrorCodeEnum {
  DEPOSIT_PROVIDER_NO_AVAILABLE = 100,
  DEPOSIT_PROVIDER_CHANNEL_NOT_FOUND = 101,
  DEPOSIT_PROVIDER_USDT_PROTO_NOT_FOUND = 102,
  DEPOSIT_PROVIDER_ORDER_NOT_CREATED = 103,
  DEPOSIT_PROVIDER_USDT_RATE_NOT_FOUND = 104,
  DEPOSIT_PROVIDER_FAILED = 105,
  DEPOSIT_PROVIDER_FAILED_INTERNAL = 106,
  WITHDRAW_IP_INVALID = 107,
  WITHDRAW_SIGN_INVALID = 108,
  WITHDRAW_PROVIDER_NO_AVAILABLE = 109,
  WITHDRAW_PROVIDER_CHANNEL_NOT_FOUND = 110,
  WITHDRAW_PROVIDER_FAILED_INTERNAL = 111,
  WITHDRAW_PROVIDER_ORDER_NOT_CREATED = 112,
  WITHDRAW_PROVIDER_BANK_NOT_SUPPORTED = 113,
  WITHDRAW_PROVIDER_FAILED = 114,
  WITHDRAW_PROVIDER_USDT_RATE_NOT_FOUND = 115,
  WITHDRAW_PROVIDER_ERROR_UNKNOWN = 998,
  DEPOSIT_PROVIDER_ERROR_UNKNOWN = 999,
}